@@ -1,6 +1,6 @@
 {
   "name": "react-grid-layout",
-  "version": "0.10.8",
+  "version": "0.12.7",
   "description": "A draggable and resizable grid layout with responsive breakpoints, for React.",
   "main": "index.js",
   "scripts": {
@@ -34,12 +34,12 @@
   "homepage": "https://github.com/STRML/react-grid-layout",
   "dependencies": {
     "lodash.isequal": "^4.0.0",
-    "react-draggable": "^1.3.1",
-    "react-resizable": "^1.2.0"
+    "react-draggable": "^2.1.1",
+    "react-resizable": "^1.4.0"
   },
   "peerDependencies": {
-    "react": "^0.14.3",
-    "react-dom": "^0.14.3"
+    "react": "^0.14.3 || ^15.0.0",
+    "react-dom": "^0.14.3 || ^15.0.0"
   },
   "devDependencies": {
     "babel-cli": "^6.5.1",
@@ -47,6 +47,9 @@
     "babel-eslint": "^4.1.8",
     "babel-loader": "^6.x",
     "babel-plugin-react-transform": "^2.0.0",
+    "babel-plugin-transform-flow-comments": "^6.7.0",
+    "babel-plugin-transform-react-constant-elements": "^6.5.0",
+    "babel-plugin-transform-react-inline-elements": "^6.6.5",
     "babel-plugin-typecheck": "^3.6.1",
     "babel-preset-es2015": "^6.5.0",
     "babel-preset-es2015-loose": "^7.0.0",
@@ -62,8 +65,8 @@
     "lodash": "^4.3.0",
     "pre-commit": "^1.1.2",
     "precommit-hook": "^3.0.0",
-    "react": "^0.14.7",
-    "react-dom": "^0.14.7",
+    "react": "^15.0.0",
+    "react-dom": "^15.0.0",
     "react-hot-loader": "^1.3.0",
     "react-transform-hmr": "^1.0.2",
     "style-loader": "^0.13.0",
