@@ -1,32 +1,30 @@
-'use strict';
-var React = require('react');
-var PureRenderMixin = require('react/lib/ReactComponentWithPureRenderMixin');
-var _ = require('lodash');
-var WidthProvider = require('react-grid-layout').WidthProvider;
-var ResponsiveReactGridLayout = require('react-grid-layout').Responsive;
-ResponsiveReactGridLayout = WidthProvider(ResponsiveReactGridLayout);
+import React from 'react';
+import _ from 'lodash';
+import {Responsive, WidthProvider} from 'react-grid-layout';
+const ResponsiveReactGridLayout = WidthProvider(Responsive);
 
-var BasicLayout = React.createClass({
-  mixins: [PureRenderMixin],
+class ShowcaseLayout extends React.Component {
 
-  propTypes: {
+  static propTypes = {
     onLayoutChange: React.PropTypes.func.isRequired
-  },
+  };
 
-  getDefaultProps() {
-    return {
-      className: "layout",
-      rowHeight: 30,
-      cols: {lg: 12, md: 10, sm: 6, xs: 4, xxs: 2}
-    };
-  },
+  static defaultProps = {
+    className: "layout",
+    rowHeight: 30,
+    cols: {lg: 12, md: 10, sm: 6, xs: 4, xxs: 2},
+    initialLayout: generateLayout()
+  };
 
-  getInitialState() {
-    return {
-      layouts: {lg: this.generateLayout()},
-      currentBreakpoint: 'lg'
-    };
-  },
+  state = {
+    currentBreakpoint: 'lg',
+    mounted: false,
+    layouts: {lg: this.props.initialLayout},
+  };
+
+  componentDidMount() {
+    this.setState({mounted: true});
+  }
 
   generateDOM() {
     return _.map(this.state.layouts.lg, function (l, i) {
@@ -38,60 +36,62 @@ var BasicLayout = React.createClass({
           }
         </div>);
     });
-  },
+  }
 
-  generateLayout() {
-    var p = this.props;
-    return _.map(_.range(0, 25), function (item, i) {
-      var y = _.result(p, 'y') || Math.ceil(Math.random() * 4) + 1;
-      return {
-        x: _.random(0, 5) * 2 % 12,
-        y: Math.floor(i / 6) * y,
-        w: 2,
-        h: y,
-        i: i.toString(),
-        static: Math.random() < 0.05
-      };
-    });
-  },
-
-  onBreakpointChange(breakpoint) {
+  onBreakpointChange = (breakpoint) => {
     this.setState({
       currentBreakpoint: breakpoint
     });
-  },
+  };
 
-  onLayoutChange(layout) {
-    this.props.onLayoutChange(layout);
-  },
+  onLayoutChange = (layout, layouts) => {
+    this.props.onLayoutChange(layout, layouts);
+  };
 
-  onNewLayout() {
+  onNewLayout = () => {
     this.setState({
-      layouts: {lg: this.generateLayout()}
+      layouts: {lg: generateLayout()}
     });
-  },
+  };
 
   render() {
     return (
       <div>
-        <div>Current Breakpoint: {this.state.currentBreakpoint} ({this.props.cols[this.state.currentBreakpoint]}
-          columns)
+        <div>Current Breakpoint: {this.state.currentBreakpoint} ({this.props.cols[this.state.currentBreakpoint]} columns)
         </div>
         <button onClick={this.onNewLayout}>Generate New Layout</button>
         <ResponsiveReactGridLayout
+          {...this.props}
           layouts={this.state.layouts}
           onBreakpointChange={this.onBreakpointChange}
           onLayoutChange={this.onLayoutChange}
-          useCSSTransforms={true}
-          {...this.props}>
+          // WidthProvider option
+          measureBeforeMount={false}
+          // I like to have it animate on mount. If you don't, delete `useCSSTransforms` (it's default `true`)
+          // and set `measureBeforeMount={true}`.
+          useCSSTransforms={this.state.mounted}>
           {this.generateDOM()}
         </ResponsiveReactGridLayout>
       </div>
     );
   }
-});
+}
 
-module.exports = BasicLayout;
+module.exports = ShowcaseLayout;
+
+function generateLayout() {
+  return _.map(_.range(0, 25), function (item, i) {
+    var y = Math.ceil(Math.random() * 4) + 1;
+    return {
+      x: _.random(0, 5) * 2 % 12,
+      y: Math.floor(i / 6) * y,
+      w: 2,
+      h: y,
+      i: i.toString(),
+      static: Math.random() < 0.05
+    };
+  });
+}
 
 if (require.main === module) {
   require('../test-hook.jsx')(module.exports);
