@@ -1,6 +1,6 @@
 // @flow
 
-import {compact, correctBounds} from './utils';
+import {cloneLayout, compact, correctBounds} from './utils';
 
 import type {Layout} from './utils';
 export type ResponsiveLayout = {lg?: Layout, md?: Layout, sm?: Layout, xs?: Layout, xxs?: Layout};
@@ -15,10 +15,10 @@ type Breakpoints = {lg?: number, md?: number, sm?: number, xs?: number, xxs?: nu
  * @return {String}       Highest breakpoint that is less than width.
  */
 export function getBreakpointFromWidth(breakpoints: Breakpoints, width: number): Breakpoint {
-  let sorted = sortBreakpoints(breakpoints);
+  const sorted = sortBreakpoints(breakpoints);
   let matching = sorted[0];
   for (let i = 1, len = sorted.length; i < len; i++) {
-    let breakpointName = sorted[i];
+    const breakpointName = sorted[i];
     if (width > breakpoints[breakpointName]) matching = breakpointName;
   }
   return matching;
@@ -56,19 +56,19 @@ export function findOrGenerateResponsiveLayout(layouts: ResponsiveLayout, breakp
                                                breakpoint: Breakpoint, lastBreakpoint: Breakpoint,
                                                cols: number, verticalCompact: boolean): Layout {
   // If it already exists, just return it.
-  if (layouts[breakpoint]) return layouts[breakpoint];
+  if (layouts[breakpoint]) return cloneLayout(layouts[breakpoint]);
   // Find or generate the next layout
   let layout = layouts[lastBreakpoint];
-  let breakpointsSorted = sortBreakpoints(breakpoints);
-  let breakpointsAbove = breakpointsSorted.slice(breakpointsSorted.indexOf(breakpoint));
+  const breakpointsSorted = sortBreakpoints(breakpoints);
+  const breakpointsAbove = breakpointsSorted.slice(breakpointsSorted.indexOf(breakpoint));
   for (let i = 0, len = breakpointsAbove.length; i < len; i++) {
-    let b = breakpointsAbove[i];
+    const b = breakpointsAbove[i];
     if (layouts[b]) {
       layout = layouts[b];
       break;
     }
   }
-  layout = JSON.parse(JSON.stringify(layout || [])); // clone layout so we don't modify existing items
+  layout = cloneLayout(layout || []); // clone layout so we don't modify existing items
   return compact(correctBounds(layout, {cols: cols}), verticalCompact);
 }
 
@@ -80,7 +80,7 @@ export function findOrGenerateResponsiveLayout(layouts: ResponsiveLayout, breakp
  * @return {Array}              Sorted breakpoints.
  */
 export function sortBreakpoints(breakpoints: Breakpoints): Array<Breakpoint> {
-  let keys: Array<string> = Object.keys(breakpoints);
+  const keys: Array<string> = Object.keys(breakpoints);
   return keys.sort(function(a, b) {
     return breakpoints[a] - breakpoints[b];
   });
