@@ -1,11 +1,21 @@
 // @flow
-export type LayoutItem = {w: number, h: number, x: number, y: number, i: string, placeholder?: boolean, moved?: boolean};
+export type LayoutItemRequired = {w: number, h: number, x: number, y: number, i: string};
+export type LayoutItem = LayoutItemRequired &
+                         {minW?: number, minH?: number, maxW?: number, maxH?: number,
+                          moved?: boolean, static?: boolean,
+                          isDraggable?: ?boolean, isResizable?: ?boolean};
 export type Layout = Array<LayoutItem>;
 export type Position = {left: number, top: number, width: number, height: number};
-export type CorePosition = {deltaX: number, deltaY: number, position: {left: number, top: number}};
+export type DragCallbackData = {
+  node: HTMLElement,
+  x: number, y: number,
+  deltaX: number, deltaY: number,
+  lastX: number, lastY: number
+};
+export type DragEvent = {e: Event} & DragCallbackData;
 export type Size = {width: number, height: number};
-export type DragEvent = {e: Event, node: HTMLElement, position: Position};
 export type ResizeEvent = {e: Event, node: HTMLElement, size: Size};
+import type React from 'react';
 
 const isProduction = process.env.NODE_ENV === 'production';
 
@@ -24,15 +34,23 @@ export function bottom(layout: Layout): number {
   return max;
 }
 
-/**
- * Clones a shallow object.
- * TODO: This could be made a lot faster if we had a clone method per datatype,
- * and just copied the properties over to a new object - this way V8 can optimize far better.
- * @param  {Object} obj Object to clone.
- * @return {Object}   Cloned object.
- */
-export function clone(obj: Object): Object {
-  return {...obj};
+export function cloneLayout(layout: Layout): Layout {
+  const newLayout = Array(layout.length);
+  for (let i = 0, len = layout.length; i < len; i++) {
+    newLayout[i] = cloneLayoutItem(layout[i]);
+  }
+  return newLayout;
+}
+
+// Fast path to cloning, since this is monomorphic
+export function cloneLayoutItem(layoutItem: LayoutItem): LayoutItem {
+  return {
+    w: layoutItem.w, h: layoutItem.h, x: layoutItem.x, y: layoutItem.y, i: layoutItem.i,
+    minW: layoutItem.minW, maxW: layoutItem.maxW, minH: layoutItem.minH, maxH: layoutItem.maxH,
+    moved: Boolean(layoutItem.moved), static: Boolean(layoutItem.static),
+    // These can be null
+    isDraggable: layoutItem.isDraggable, isResizable: layoutItem.isResizable
+  };
 }
 
 /**
@@ -76,9 +94,11 @@ export function collides(l1: LayoutItem, l2: LayoutItem): boolean {
  */
 export function compact(layout: Layout, verticalCompact: boolean): Layout {
   // Statics go in the compareWith array right away so items flow around them.
-  let compareWith = getStatics(layout), out = [];
+  const compareWith = getStatics(layout);
   // We go through the items by row and column.
-  let sorted = sortLayoutItemsByRowCol(layout);
+  const sorted = sortLayoutItemsByRowCol(layout);
+  // Holding for new items.
+  const out = Array(layout.length);
 
   for (let i = 0, len = sorted.length; i < len; i++) {
     let l = sorted[i];
@@ -128,9 +148,9 @@ export function compactItem(compareWith: Layout, l: LayoutItem, verticalCompact:
  * @param  {Number} bounds Number of columns.
  */
 export function correctBounds(layout: Layout, bounds: {cols: number}): Layout {
-  let collidesWith = getStatics(layout);
+  const collidesWith = getStatics(layout);
   for (let i = 0, len = layout.length; i < len; i++) {
-    let l = layout[i];
+    const l = layout[i];
     // Overflows right
     if (l.x + l.w > bounds.cols) l.x = bounds.cols - l.w;
     // Overflows left
@@ -146,7 +166,6 @@ export function correctBounds(layout: Layout, bounds: {cols: number}): Layout {
         l.y++;
       }
     }
-
   }
   return layout;
 }
@@ -179,11 +198,7 @@ export function getFirstCollision(layout: Layout, layoutItem: LayoutItem): ?Layo
 }
 
 export function getAllCollisions(layout: Layout, layoutItem: LayoutItem): Array<LayoutItem> {
-  let out = [];
-  for (let i = 0, len = layout.length; i < len; i++) {
-    if (collides(layout[i], layoutItem)) out.push(layout[i]);
-  }
-  return out;
+  return layout.filter((l) => collides(l, layoutItem));
 }
 
 /**
@@ -192,11 +207,7 @@ export function getAllCollisions(layout: Layout, layoutItem: LayoutItem): Array<
  * @return {Array}        Array of static layout items..
  */
 export function getStatics(layout: Layout): Array<LayoutItem> {
-  let out = [];
-  for (let i = 0, len = layout.length; i < len; i++) {
-    if (layout[i].static) out.push(layout[i]);
-  }
-  return out;
+  return layout.filter((l) => l.static);
 }
 
 /**
@@ -213,7 +224,7 @@ export function moveElement(layout: Layout, l: LayoutItem, x: ?number, y: ?numbe
   // Short-circuit if nothing to do.
   if (l.y === y && l.x === x) return layout;
 
-  let movingUp = y && l.y > y;
+  const movingUp = y && l.y > y;
   // This is quite a bit faster than extending the object
   if (typeof x === 'number') l.x = x;
   if (typeof y === 'number') l.y = y;
@@ -225,11 +236,11 @@ export function moveElement(layout: Layout, l: LayoutItem, x: ?number, y: ?numbe
   // nearest collision.
   let sorted = sortLayoutItemsByRowCol(layout);
   if (movingUp) sorted = sorted.reverse();
-  let collisions = getAllCollisions(sorted, l);
+  const collisions = getAllCollisions(sorted, l);
 
   // Move each item that collides away from this element.
   for (let i = 0, len = collisions.length; i < len; i++) {
-    let collision = collisions[i];
+    const collision = collisions[i];
     // console.log('resolving collision between', l.i, 'at', l.y, 'and', collision.i, 'at', collision.y);
 
     // Short circuit so we can't infinite loop
@@ -267,7 +278,7 @@ export function moveElementAwayFromCollision(layout: Layout, collidesWith: Layou
   // unwanted swapping behavior.
   if (isUserAction) {
     // Make a mock item so we don't modify the item here, only modify in moveElement.
-    let fakeItem = {
+    const fakeItem: LayoutItem = {
       x: itemToMove.x,
       y: itemToMove.y,
       w: itemToMove.w,
@@ -295,16 +306,29 @@ export function perc(num: number): string {
   return num * 100 + '%';
 }
 
-export function setTransform(style: Object, coords: [number, number]): Object {
+export function setTransform({top, left, width, height}: Position): Object {
   // Replace unitless items with px
-  let x = ('' + coords[0]).replace(/(\d)$/, '$1px');
-  let y = ('' + coords[1]).replace(/(\d)$/, '$1px');
-  style.transform = "translate(" + x + "," + y + ")";
-  style.WebkitTransform = "translate(" + x + "," + y + ")";
-  style.MozTransform = "translate(" + x + "," + y + ")";
-  style.msTransform = "translate(" + x + "," + y + ")";
-  style.OTransform = "translate(" + x + "," + y + ")";
-  return style;
+  const translate = `translate(${left}px,${top}px)`;
+  return {
+    transform: translate,
+    WebkitTransform: translate,
+    MozTransform: translate,
+    msTransform: translate,
+    OTransform: translate,
+    width: `${width}px`,
+    height: `${height}px`,
+    position: 'absolute'
+  };
+}
+
+export function setTopLeft({top, left, width, height}: Position): Object {
+  return {
+    top: `${top}px`,
+    left: `${left}px`,
+    width: `${width}px`,
+    height: `${height}px`,
+    position: 'absolute'
+  };
 }
 
 /**
@@ -331,7 +355,7 @@ export function sortLayoutItemsByRowCol(layout: Layout): Layout {
  * @param  {Boolean} verticalCompact Whether or not to compact the layout vertically.
  * @return {Array}                Working layout.
  */
-export function synchronizeLayoutWithChildren(initialLayout: Layout, children: Array<ReactElement>|ReactElement,
+export function synchronizeLayoutWithChildren(initialLayout: Layout, children: Array<React.Element<any>>|React.Element<any>,
                                               cols: number, verticalCompact: boolean): Layout {
   // ensure 'children' is always an array
   if (!Array.isArray(children)) {
@@ -340,33 +364,41 @@ export function synchronizeLayoutWithChildren(initialLayout: Layout, children: A
   initialLayout = initialLayout || [];
 
   // Generate one layout item per child.
-  let layout = [];
+  let layout: Layout = [];
   for (let i = 0, len = children.length; i < len; i++) {
-    let child = children[i];
+    let newItem;
+    const child = children[i];
+
     // Don't overwrite if it already exists.
-    let exists = getLayoutItem(initialLayout, child.key || "1" /* FIXME satisfies Flow */);
+    const exists = getLayoutItem(initialLayout, child.key || "1" /* FIXME satisfies Flow */);
     if (exists) {
-      layout.push(exists);
-      continue;
-    }
-    // New item: attempt to use a layout item from the child, if it exists.
-    let g = child.props._grid;
-    if (g) {
-      if (!isProduction) {
-        validateLayout([g], 'ReactGridLayout.children');
-      }
-      // Validated; add it to the layout. Bottom 'y' possible is the bottom of the layout.
-      // This allows you to do nice stuff like specify {y: Infinity}
-      if (verticalCompact) {
-        // TODO create cloneLayoutItem for speed
-        layout.push({...g, y: Math.min(bottom(layout), g.y), i: child.key});
-      } else {
-        layout.push({...g, y: g.y, i: child.key});
-      }
+      newItem = cloneLayoutItem(exists);
     } else {
+      if (process.env.NODE_ENV !== 'production' && child.props._grid) {
+        console.warn('`_grid` properties on children have been deprecated as of React 15.2. ' + // eslint-disable-line
+          'Please use `data-grid` or add your properties directly to the `layout`.');
+      }
+      const g = child.props['data-grid'] || child.props._grid;
+
+      // Hey, this item has a data-grid property, use it.
+      if (g) {
+        if (!isProduction) {
+          validateLayout([g], 'ReactGridLayout.children');
+        }
+        // Validated; add it to the layout. Bottom 'y' possible is the bottom of the layout.
+        // This allows you to do nice stuff like specify {y: Infinity}
+        if (verticalCompact) {
+          newItem = cloneLayoutItem({...g, y: Math.min(bottom(layout), g.y), i: child.key});
+        } else {
+          newItem = cloneLayoutItem({...g, y: g.y, i: child.key});
+        }
+      }
       // Nothing provided: ensure this is added to the bottom
-      layout.push({w: 1, h: 1, x: 0, y: bottom(layout), i: child.key || "1"});
+      else {
+        newItem = cloneLayoutItem({w: 1, h: 1, x: 0, y: bottom(layout), i: child.key || "1"});
+      }
     }
+    layout[i] = newItem;
   }
 
   // Correct the layout.
@@ -385,10 +417,10 @@ export function synchronizeLayoutWithChildren(initialLayout: Layout, children: A
  */
 export function validateLayout(layout: Layout, contextName: string): void {
   contextName = contextName || "Layout";
-  let subProps = ['x', 'y', 'w', 'h'];
+  const subProps = ['x', 'y', 'w', 'h'];
   if (!Array.isArray(layout)) throw new Error(contextName + " must be an array!");
   for (let i = 0, len = layout.length; i < len; i++) {
-    let item = layout[i];
+    const item = layout[i];
     for (let j = 0; j < subProps.length; j++) {
       if (typeof item[subProps[j]] !== 'number') {
         throw new Error('ReactGridLayout: ' + contextName + '[' + i + '].' + subProps[j] + ' must be a number!');
