@@ -4,6 +4,8 @@ var PureRenderMixin = require('react/lib/ReactComponentWithPureRenderMixin');
 var WidthProvider = require('react-grid-layout').WidthProvider;
 var ReactGridLayout = require('react-grid-layout');
 ReactGridLayout = WidthProvider(ReactGridLayout);
+
+const originalLayout = getFromLS('layout') || [];
 /**
  * This layout demonstrates how to sync to localstorage.
  */
@@ -19,31 +21,21 @@ var LocalStorageLayout = React.createClass({
   },
 
   getInitialState() {
-    var ls = {};
-    if (global.localStorage) {
-      try {
-        ls = JSON.parse(global.localStorage.getItem('rgl-7')) || {};
-      } catch(e) {/*ignore*/}
-    }
-    return {layout: ls.layout || []};
+    return {
+      layout: JSON.parse(JSON.stringify(originalLayout))
+    };
   },
 
   resetLayout() {
-    this.setState({layout: []});
-  },
-
-  _saveToLocalStorage() {
-    if (global.localStorage) {
-      global.localStorage.setItem('rgl-7', JSON.stringify({
-        layout: this.state.layout
-      }));
-    }
+    this.setState({
+      layout: []
+    });
   },
 
   onLayoutChange(layout) {
     /*eslint no-console: 0*/
-    console.log('layout changed', layout);
-    this._saveToLocalStorage();
+    saveToLS('layout', layout);
+    this.setState({layout});
     this.props.onLayoutChange(layout); // updates status display
   },
 
@@ -52,19 +44,38 @@ var LocalStorageLayout = React.createClass({
       <div>
         <button onClick={this.resetLayout}>Reset Layout</button>
         <ReactGridLayout
+            ref="rgl"
             {...this.props}
             layout={this.state.layout}
             onLayoutChange={this.onLayoutChange}>
-          <div key="1" _grid={{w: 2, h: 3, x: 0, y: 0}}><span className="text">1</span></div>
-          <div key="2" _grid={{w: 2, h: 3, x: 2, y: 0}}><span className="text">2</span></div>
-          <div key="3" _grid={{w: 2, h: 3, x: 4, y: 0}}><span className="text">3</span></div>
-          <div key="4" _grid={{w: 2, h: 3, x: 6, y: 0}}><span className="text">4</span></div>
-          <div key="5" _grid={{w: 2, h: 3, x: 8, y: 0}}><span className="text">5</span></div>
+          <div key="1" data-grid={{w: 2, h: 3, x: 0, y: 0}}><span className="text">1</span></div>
+          <div key="2" data-grid={{w: 2, h: 3, x: 2, y: 0}}><span className="text">2</span></div>
+          <div key="3" data-grid={{w: 2, h: 3, x: 4, y: 0}}><span className="text">3</span></div>
+          <div key="4" data-grid={{w: 2, h: 3, x: 6, y: 0}}><span className="text">4</span></div>
+          <div key="5" data-grid={{w: 2, h: 3, x: 8, y: 0}}><span className="text">5</span></div>
         </ReactGridLayout>
       </div>
     );
   }
 });
+
+function getFromLS(key) {
+  let ls = {};
+  if (global.localStorage) {
+    try {
+      ls = JSON.parse(global.localStorage.getItem('rgl-7')) || {};
+    } catch(e) {/*Ignore*/}
+  }
+  return ls[key];
+}
+
+function saveToLS(key, value) {
+  if (global.localStorage) {
+    global.localStorage.setItem('rgl-7', JSON.stringify({
+      [key]: value
+    }));
+  }
+}
 
 module.exports = LocalStorageLayout;
 
